fix: decklink vanc line datatype

+++ src/modules/decklink/consumer/vanc.h
@@ -28,9 +28,9 @@ namespace caspar { namespace decklink {
 
 struct vanc_packet
 {
-    uint8_t              did;
-    uint8_t              sdid;
-    uint32_t             line_number;
+    uint8_t               did;
+    uint8_t               sdid;
+    uint32_t              line_number;
     std::vector<uint16_t> data;
 };
 
@@ -57,8 +57,8 @@ class decklink_vanc
 };
 
 std::shared_ptr<decklink_vanc_strategy>
-create_op47_strategy(uint8_t line_number, uint8_t line_number_2, const std::wstring& dummy_header);
-std::shared_ptr<decklink_vanc_strategy> create_scte104_strategy(uint8_t line_number);
+create_op47_strategy(uint32_t line_number, uint32_t line_number_2, const std::wstring& dummy_header);
+std::shared_ptr<decklink_vanc_strategy> create_scte104_strategy(uint32_t line_number);
 
 std::shared_ptr<decklink_vanc> create_vanc(const vanc_configuration& config);
